==> notebooks/01_data_merge_2024.ipynb <==
import pandas as pd
import glob
import os

# Pfad zu den Daten definieren
path = "../data/raw/weekly-top-songs-global-2024/*.csv"
all_files = glob.glob(path)

all_dfs_24 = []

for filename in all_files: 
    # CSV einlesen
    df = pd.read_csv(filename, index_col=None, header=0)

    # Datum aus Dateinamen extrahieren
    # Logik: regional-global-weekly-yyyy-mm-dd.csv -> splitte bei '-' und nimm die letzten drei 
    date_str = filename.split('-')[-3:] # ['2024', '01', '04.csv']
    date_str = "-".join(date_str).replace('.csv', '') # 2024-01-04

    # Neue Spalte hinzufügen
    df['chart_week'] = pd.to_datetime(date_str)

    all_dfs_24.append(df)

# Alle DataFrames zu einem großen Set zusammenfügen
df_all = pd.concat(all_dfs_24, axis=0, ignore_index=True)

# Erste Kontrolle
print(f"Datensatz geladen: {df_all.shape[0]} Zeilen aus {len(all_files)} Dateien.")
df_all.head()

# Definieren der relevanten Spalten
keep_columns = [
    'chart_week', 'rank', 'uri', 'artist_names', 
    'track_name', 'peak_rank', 'previous_rank', 
    'weeks_on_chart', 'streams'
]

# Filtern des DataFrames
df_clean_2024 = df_all[keep_columns].copy()

display(df_clean_2024.head())
display(df_clean_2024.info())

# Als CSV speichern
# Zielverzeichnis definieren
output_dir = '../data/interim'
output_file = os.path.join(output_dir, 'df_cleaned_2024.csv')

# Ordner erstellen, falls er noch nicht existiert
if not os.path.exists(output_dir):
    os.makedirs(output_dir)
    print(f"Ordner erstellt: {output_dir}")

# DataFrame als CSV speichern
df_clean_2024.to_csv(output_file, index=False, encoding='utf-8')
print(f"Datei erfolgreich gespeichert unter: {output_file}")
