==> notebooks/05_extraction_unique_entities.ipynb <==
import pandas as pd
import os

# Lade deine bereinigten Charts aus Woche 1
df_charts = pd.read_csv("../data/processed/df_cleaned_full.csv") 
print("Daten erfolgreich geladen.")

# Eindeutige Künstler (für Genre & Follower)
unique_artists = df_charts['artist_names'].unique()

# Eindeutige Kombinationen aus Künstler und Track extrahieren
# Beides nötig, um später IDs eindeutig zuordnen zu können
unique_tracks = df_charts[['artist_names', 'track_name']].drop_duplicates().reset_index(drop=True)

print(f"Gesamtanzahl Chart-Einträge: {len(df_charts)}")
print(f"Anzahl eindeutiger Tracks: {len(unique_tracks)}")
print(f"Anzahl eindeutiger Künstler: {len(unique_artists)}")

unique_tracks.head()

# Als CSV speichern
# Zielverzeichnis definieren
output_dir = '../data/interim'
output_file = os.path.join(output_dir, 'unique_tracks_to_enrich.csv')

# Ordner erstellen, falls er noch nicht existiert
if not os.path.exists(output_dir):
    os.makedirs(output_dir)
    print(f"Ordner erstellt: {output_dir}")

# Als Gesamtdatei speichern
unique_tracks.to_csv("../data/interim/unique_tracks_to_enrich.csv", index=False, encoding='utf-8')
print(f"Datei erfolgreich gespeichert unter: {output_file}")
